requirements.txt
Flask
fal-client
Pillow
requests
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTful
python-dotenv
asyncio
Werkzeug
flask-login
psycopg2-binary
flask_sqlalchemy
gunicorn